# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Team Titans - Free Lunch API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/organization/create": {
            "put": {
                "tags": [
                    "Organization"
                ],
                "summary": "Create a new organization",
                "operationId": "37c9e238c8b894be5d7c84ee4aa74f39",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Example Organization",
                                    "description": "A sample organization"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Create a new organization",
                                        "value": {
                                            "message": "success",
                                            "statusCode": 200,
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Create a new organization",
                                        "value": {
                                            "message": "You are not authorized to perform this action!"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Organization"
                ],
                "summary": "Create Organization",
                "operationId": "cc783a78127b97001fc10c30994efd67",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "organization_name": {
                                        "type": "string"
                                    },
                                    "lunch_price": {
                                        "type": "integer"
                                    },
                                    "currency_code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "organization_name": "Example Organization",
                                    "lunch_price": 1000,
                                    "currency_code": "123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Organization create",
                                        "value": {
                                            "status": "success",
                                            "message": "Organization Created Successfully",
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE_ENTITY",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Organization create",
                                        "value": {
                                            "status": "failed",
                                            "message": "You already belong to a company",
                                            "data": null
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/organization/staff/signup": {
            "post": {
                "tags": [
                    "Organization"
                ],
                "summary": "Create a user within an organization",
                "operationId": "620e70f3dfc4b9a6177c4c60d520aab8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "otp_token": {
                                        "type": "string"
                                    },
                                    "is_admin": {
                                        "type": "boolean"
                                    },
                                    "org_id": {
                                        "type": "integer"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "profile_pic": {
                                        "type": "file",
                                        "example": "picture_url.jpg"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "first_name": "John",
                                    "last_name": "Doe",
                                    "email": "john.doe@example.com",
                                    "otp_token": "123456",
                                    "is_admin": false,
                                    "org_id": 1,
                                    "phone": "1234567890",
                                    "profile_pic": "example.jpg"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Create a user within an organization using an invitation token",
                                        "value": {
                                            "status_code": 201,
                                            "message": "User signed up successfully",
                                            "status": "success",
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Create a user within an organization using an invitation token",
                                        "value": {
                                            "status_code": 401,
                                            "status": "failed",
                                            "message": "Authentication failed"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/organization/lunch-price": {
            "patch": {
                "tags": [
                    "Organization"
                ],
                "summary": "Update lunch price",
                "operationId": "c5c1cf536c432da7e6c0861294d9adb0",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lunch_price": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "lunch_price": 1000
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Update lunch price",
                                        "value": {
                                            "status": "success",
                                            "message": "Lunch price updated successfully",
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Update lunch price",
                                        "value": {
                                            "status": "failed",
                                            "message": "You are not authorized to perform this action",
                                            "data": null
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE_ENTITY",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Update lunch price",
                                        "value": {
                                            "status": "failed",
                                            "message": "Error updating lunch price",
                                            "data": null
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/organization": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Get a list of organizations not deleted",
                "operationId": "2b2c84a4a824831ef20e7a0ba7bca981",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Get list of organizations not deleted",
                                        "value": {
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}